Move redux middlware into its own folder
