+++ src/components/services/Services.js
@@ -13,9 +13,9 @@ import { loadDeviceReferral } from '../../actions/DeviceReferralActions.js'
 import { rootReducer } from '../../reducers/RootReducer.js'
 import { type Action } from '../../types/reduxActions.js'
 import { type Dispatch, type RootState } from '../../types/reduxTypes.js'
-import errorAlert from '../../util/errorAlert.js'
-import loginStatusChecker from '../../util/loginStatusChecker.js'
-import perfLogger from '../../util/perfLogger.js'
+import errorAlert from '../../util/middleware/errorAlert.js'
+import loginStatusChecker from '../../util/middleware/loginStatusChecker.js'
+import perfLogger from '../../util/middleware/perfLogger.js'
 import { ModalProvider } from '../common/ModalProvider.js'
 import { Main } from '../Main.ui.js'
 import { Airship } from './AirshipInstance.js'
