+++ src/util/middleware/perfLogger.js
@@ -2,7 +2,7 @@
 
 import RNFS from 'react-native-fs'
 
-import type { Action, Next, Store } from '../types/reduxTypes.js'
+import type { Action, Next, Store } from '../../types/reduxTypes.js'
 
 const perfLoggerCSV = RNFS.DocumentDirectoryPath + '/perfLogger.csv'
 RNFS.writeFile(perfLoggerCSV, 'action type,start,end\n', 'utf8')
